refactor: refactor ad handling and update app version
- Removed the imports of `google_mobile_ads` and `toc_machine_trading_fe/core/ad/ad.dart` from `news.dart` and `search.dart`.
- Instead, imported `toc_machine_trading_fe/features/universal/widgets/ad.dart` in the same files.
- Removed the handling of ad loading and orientation changes from `_NewsPageState` and `_SearchFuturePageState`.
- In `news.dart` and `search.dart`, replaced the custom ad loading logic with a simple `AdBanner` widget.
- Created a new file `ad.dart` in the `universal/widgets` directory, which contains the `AdBanner` widget. This widget handles the ad loading and orientation change logic that was previously in `news.dart` and `search.dart`.
- In `search.dart`, added logic to unfocus the text field when the user taps elsewhere on the screen.
- Updated the app version in `pubspec.yaml` from `4.2.57+40117` to `4.2.58+40118`.
