+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: toc_machine_trading_fe
 description: "Status of trade agent, let user query analyzed stocks."
 publish_to: "none"
-version: 4.2.57+40117
+version: 4.2.58+40118
 
 environment:
   sdk: ">=3.2.5 <4.0.0"
